pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "goose-mentor-mode"
version = "0.1.0"
description = "AI-powered mentor extension for Goose that transforms automation into guided learning"
authors = [
    {name = "Joe Euston", email = "joe@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Education",
]
keywords = ["goose", "ai", "mentor", "learning", "development"]

dependencies = [
    # Core dependencies for mentor functionality
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
goose = [
    "goose-ai>=0.9.0",
    "rich>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/joeeuston-dev/goose-mentor-mode"
Repository = "https://github.com/joeeuston-dev/goose-mentor-mode"
Documentation = "https://github.com/joeeuston-dev/goose-mentor-mode/blob/main/README.md"
Issues = "https://github.com/joeeuston-dev/goose-mentor-mode/issues"

[project.entry-points."goose.toolkit"]
mentor = "goose_mentor_mode:MentorToolkit"

[tool.hatch.build.targets.wheel]
packages = ["src/goose_mentor_mode"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
